PK Long->Integer로 변경

+++ src/main/java/kr/ac/bc/comon/springboot/endpoint/field/FieldController.java
@@ -16,7 +16,7 @@ public class FieldController {
     private final FieldService fieldService;
 
     @GetMapping("/{id}")
-    public FieldResponseDto findById(@PathVariable Long id) {
+    public FieldResponseDto findById(@PathVariable Integer id) {
         return fieldService.findById(id);
     }
