+++ src/test/java/kr/ac/bc/comon/springboot/endpoint/user/UserControllerTest.java
@@ -53,7 +53,7 @@ public class UserControllerTest {
         String url = "http://localhost:" + port + "/user/join";
 
         //when
-        ResponseEntity<Long> responseEntity = restTemplate.postForEntity(url, requestDto, Long.class);
+        ResponseEntity<Integer> responseEntity = restTemplate.postForEntity(url, requestDto, Integer.class);
 
         //then
         assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
